+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## v4.0.0-alpha1
+
+- Upgrade `cosmos-sdk` to `v0.45.3` and `ibc-go` to `v3.0.0`. #717 #725 #728 #745 #750
+- Mark gRPC and CLI of `x/supply` as deprecated. #724
+- Enable interchain accounts. #730 #732 #733
+
+*March 1, 2022*
+
+## v3.3.4
+
+This version is identical to v3.3.3, but 
+
+- Bump `cosmos-sdk`, `ibc-go` and `tendermint` to latest minor versions. #711
+- Upgrade Swagger UI version for security fixes. #704
+
 *December 6, 2021*
 
 ## v3.3.3